requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
et-xmlfile==1.1.0
fonttools==4.39.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.38.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.59.0
httplib2==0.21.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
moviepy==1.0.3
multidict==6.0.4
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
proglog==0.1.10
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydub==0.25.1
pygame==2.2.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
SpeechRecognition==3.9.0
statsmodels==0.13.5
threadpoolctl==3.1.0
tifffile==2023.3.21
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whisper==1.1.10
xlrd==2.0.1
yarl==1.8.2
